requirements.txt
aniso8601==8.1.0
apig-wsgi==2.0.1
appdirs==1.4.4
attrs==20.3.0
backports.functools-lru-cache==1.6.1
blinker==1.4
boto3==1.14.49
botocore==1.17.63
bson
caniusepython3==7.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
docutils==0.15.2
filelock==3.0.12
Flask==1.0.3
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
flask-restplus==0.12.1
future==0.17.1
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.3.1
logzero==1.6.3
MarkupSafe==1.1.1
marshmallow==3.12.1
matplotlib==2.2.4
networkx==2.2
numpy
packaging==20.8
pandas==0.24.2
pbr==5.4.5
pymongo
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
scipy
sentry-sdk==0.9.0
six==1.15.0
stevedore==3.2.0
tqdm==4.60.0
urllib3
virtualenv==20.2.2
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==0.16.1
zipp==3.1.0
